+++ jgclark.Dashboard/src/react/components/ItemRow.jsx
@@ -3,7 +3,7 @@
 // Represents a row item within a section.
 // Could be: Task, Review Item, Filtered Indicator, No Tasks left, No Projects, No Search Results.
 // Called by ItemGrid component.
-// Last updated 2025-11-23 for v2.3.0.b15, @jgclark
+// Last updated 2025-12-01 for v2.3.1, @jgclark
 //--------------------------------------------------------------------------
 
 import React, { type Node } from 'react'
@@ -29,9 +29,9 @@ type Props = {
 function ItemRow({ item, thisSection, onToggleShowAll }: Props): Node {
   const { itemType } = item
 
-  let congratsMessage = 'Nothing on this list'
+  let itemCongratsMessage = 'Nothing on this list'
   if (itemType === 'itemCongrats' && thisSection.doneCounts?.completedTasks && thisSection.doneCounts.completedTasks > 0) {
-    congratsMessage = `All ${thisSection.doneCounts.completedTasks} items completed on this list`
+    itemCongratsMessage = `All ${thisSection.doneCounts.completedTasks} items completed on this list`
   }
 
   // Deal with the different item types, defaulting to a task/checklist at the end
@@ -52,8 +52,10 @@ function ItemRow({ item, thisSection, onToggleShowAll }: Props): Node {
         />
       ) : itemType === 'filterIndicator' ? (
         <TasksFiltered item={item} onToggleShowAll={onToggleShowAll} />
+              ) : itemType === 'offerToFilter' ? (
+                <MessageOnlyItem message={item?.message ?? ''} contentClassName="messageItemRow" rowIconClassName="fa-regular fa-plus" settingsDialogAnchor={item?.settingsDialogAnchor ?? ''} />
       ) : itemType === 'itemCongrats' ? (
-        <MessageOnlyItem message={congratsMessage} contentClassName="itemCongrats" closingFAIconClassName="fa-light fa-champagne-glasses pad-left" />
+                    <MessageOnlyItem message={itemCongratsMessage} contentClassName="itemCongrats" closingFAIconClassName="fa-light fa-champagne-glasses pad-left" />
       ) : itemType === 'info' ? (
                     <MessageOnlyItem message={item?.message ?? ''} contentClassName="infoItemRow" />
       ) : (