+++ jgclark.Dashboard/src/react/components/TasksFiltered.jsx
@@ -2,7 +2,7 @@
 //--------------------------------------------------------------------------
 // Dashboard React component to show an Indicator that a Filter has been applied and so some item(s) have been hidden.
 // Called by ItemRow component
-// Last updated 2025-11-29 for v2.3.0.b16, @jgclark
+// Last updated 2025-12-01 for v2.3.2, @jgclark
 //--------------------------------------------------------------------------
 
 import React, { type Node } from 'react'
@@ -40,7 +40,7 @@ const TasksFiltered = ({ item, onToggleShowAll }: Props): Node => {
   }
 
   return (
-    <div className="sectionItemRow messageItemRow" id={item.ID}>
+    <div className="sectionItemRow messageItemRow clickTarget" id={item.ID}>
       {/* This empty span needed to mimic the StatusIcon line */}
       <span>
         <div className="itemIcon">
@@ -49,7 +49,7 @@ const TasksFiltered = ({ item, onToggleShowAll }: Props): Node => {
       </span>
       <div className="sectionItemContent sectionItem" onClick={(e) => handleLineClick(e)}>
         <span className="messageItemRow">
-          <i>{item?.para?.content || '<no content>'}</i>
+          <i>{item?.message || '<no content>'}</i>
         </span>
       </div>
     </div>